projects/project1_1/requirements.txt
# 基于LangChain的多任务问答助手 - 依赖包
langchain==0.1.17
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-core==0.1.52
tavily-python==0.3.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.3.4
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1